requirements.txt
amqp==5.0.1
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.1
django-redis==4.12.1
djangorestframework==3.11.1
idna==2.10
kombu==5.0.2
prompt-toolkit==3.0.7
psycopg2-binary==2.8.6
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5